gh-2252 Moved code to a separate file
